packages/image_utils/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "image_utils"
description = "Image Utilities"
version = "0.1.0"
authors = [
  { name = "Oghenetega Biokoro", email = "teggyg123@gmail.com" }
]

dependencies = [
  "comfy_creator", "boto3", "pillow", "numpy", "python-dotenv"
]


[project.entry-points."comfy_creator.custom_nodes"]
save_file = "image_utils.custom_nodes:SaveFile"